Refactor core to user app
